requirements.txt
azure-functions==1.21.3
numpy==2.2.1
opencv-python-headless==4.10.0.84
azure-storage-blob==12.10.0
azure-data-tables==12.5.0